added follow feature and also followers function to the model to get list of user that follow the user (followers)

+++ app/Http/Controllers/PostsController.php
@@ -36,12 +36,15 @@ class PostsController extends Controller
 
     public function destroy(Post $post)
     {
+        if (auth()->id() != $post->owner_id)  {
+            abort(403);
+        }
+
         $post->delete();
 
         if (\request()->wantsJson()) {
             return $post;
         }
-
         return redirect('/posts');
     }
 }
